refactor(backend): Unify friend and collective sub-resource bases
The friend and collective sub-resource CRUD bases were diverged copies:
the collective one wrapped primary-flag clear+write in a transaction
(race-safe), the friend one did not. Merge them into one owner-keyed
generic base (src/services/base/sub-resource.service.ts) with optional
listing and count-or-list auto-primary, and point all eight concrete
services at it. Friend sub-resources now inherit the transactional
primary-flag handling — which is the BUG-2 fix — and the duplicate base
is deleted.

Addresses DUP-1 and BUG-2 from the 2026-06-13 backend review.
