+++ apps/backend/src/services/collectives/sub-resources/phone.service.ts
@@ -10,27 +10,27 @@ import {
 } from '../../../models/queries/collective-phones.queries.js';
 import { parsePhoneType } from '../../../utils/type-guards.js';
 import {
-  CollectiveSubResourceService,
-  type CollectiveSubResourceServiceOptions,
-} from './base.service.js';
+  SubResourceService,
+  type SubResourceServiceOptions,
+} from '../../base/sub-resource.service.js';
 
 /**
  * PhoneService handles all phone-related operations for collectives.
- * Extends CollectiveSubResourceService for common CRUD operations.
+ * Extends SubResourceService for common CRUD operations.
  */
-export class CollectivePhoneService extends CollectiveSubResourceService<
+export class CollectivePhoneService extends SubResourceService<
   PhoneInput,
   Phone,
   IGetPhonesByCollectiveIdResult,
   IGetPhonesByCollectiveIdResult,
   IDeletePhoneResult
 > {
-  constructor(options: CollectiveSubResourceServiceOptions) {
+  constructor(options: SubResourceServiceOptions) {
     super(options, {
       resourceName: 'collective phone',
       hasPrimaryFlag: true,
 
-      listFn: async ({ userExternalId, collectiveExternalId }, client) => {
+      listFn: async ({ userExternalId, ownerExternalId: collectiveExternalId }, client) => {
         return getPhonesByCollectiveId.run({ userExternalId, collectiveExternalId }, client);
       },
 
@@ -43,7 +43,10 @@ export class CollectivePhoneService extends CollectiveSubResourceService<
         createdAt: row.created_at.toISOString(),
       }),
 
-      createFn: async ({ userExternalId, collectiveExternalId, input }, client) => {
+      createFn: async (
+        { userExternalId, ownerExternalId: collectiveExternalId, input },
+        client,
+      ) => {
         return createPhone.run(
           {
             userExternalId,
@@ -58,7 +61,7 @@ export class CollectivePhoneService extends CollectiveSubResourceService<
       },
 
       updateFn: async (
-        { userExternalId, collectiveExternalId, resourceExternalId, input },
+        { userExternalId, ownerExternalId: collectiveExternalId, resourceExternalId, input },
         client,
       ) => {
         return updatePhone.run(
@@ -75,7 +78,10 @@ export class CollectivePhoneService extends CollectiveSubResourceService<
         );
       },
 
-      deleteFn: async ({ userExternalId, collectiveExternalId, resourceExternalId }, client) => {
+      deleteFn: async (
+        { userExternalId, ownerExternalId: collectiveExternalId, resourceExternalId },
+        client,
+      ) => {
         return deletePhone.run(
           {
             userExternalId,
@@ -86,7 +92,7 @@ export class CollectivePhoneService extends CollectiveSubResourceService<
         );
       },
 
-      clearPrimaryFn: async ({ userExternalId, collectiveExternalId }, client) => {
+      clearPrimaryFn: async ({ userExternalId, ownerExternalId: collectiveExternalId }, client) => {
         return clearPrimaryPhone.run({ userExternalId, collectiveExternalId }, client);
       },
